Initial import of dotfiles
The .bash_profile (for interactive bash shells) loads .profile
which then loads .bashrc (for non-interactive bash shells) and
.bash_aliases.  The .mailcap is set up to help with using an html
viewer in mutt for viewing html-formatted e-mail. The .gitconfig
provides some nice starting settings, but a couple of fields have
template user information values that need to be filled in before using
it.  The .mailcap provides a viewer for html e-mail.  The .vimrc starts
by loading vim-sensible and then layers additional settings on top of
that.