Add user ssh config to make behavior more consistent between systems
Disable X11 forwarding by default, in part so that when you really want
the command-line mode of a program, it doesn't decide that there's an X
display it can use. Disable compression, since it's uncommon to be using
a link slow enough to make it worthwhile.  Set a keepalive so that
connections don't terminate due to idleness.